requirements.txt
flask==3.0.0
flask-cors==4.0.0
openai==1.3.0
pandas==2.1.3
numpy==1.26.2
scipy==1.11.4
python-dotenv==1.0.0
# Note: sqlite3 is built into Python, no need to install
# Note: Removed riskfolio-lib, now using scipy.optimize for portfolio optimization

